test: move tests of extracted methods and classes into new tests
