+++ __tests__/unit/services/client.spec.js
@@ -1,13 +1,12 @@
 import { cloneDeep } from 'lodash'
 import nock from 'nock'
-import { Identities, Managers, Transactions } from '@arkecosystem/crypto'
+import { Identities, Managers } from '@arkecosystem/crypto'
 import errorCapturer from '../__utils__/error-capturer'
 import fixtures from '../__fixtures__/services/client'
 import ClientService from '@/services/client'
 import BigNumber from '@/plugins/bignumber'
 import store from '@/store'
 import logger from 'electron-log'
-import TransactionService from '@/services/transaction'
 import WalletService from '@/services/wallet'
 
 const sessionNetwork = Object.freeze({
@@ -197,29 +196,6 @@ describe('Services > Client', () => {
 
       expect(ClientService.__watchProfile).not.toHaveBeenCalled()
       ClientService.__watchProfile = watchProfileOriginal
-    })
-  })
-
-  describe('normalizePassphrase', () => {
-    it('should normalize if provided', () => {
-      const spy = jest.spyOn(String.prototype, 'normalize')
-
-      const passphrase = client.normalizePassphrase('test')
-
-      expect(spy).toHaveBeenNthCalledWith(1, 'NFD')
-      expect(passphrase).toBe('test')
-
-      spy.mockRestore()
-    })
-
-    it('should not normalize if no passphrase', () => {
-      const spy = jest.spyOn(String.prototype, 'normalize')
-
-      client.normalizePassphrase(null)
-
-      expect(spy).not.toHaveBeenCalled()
-
-      spy.mockRestore()
     })
   })
 
@@ -1879,550 +1855,6 @@ describe('Services > Client', () => {
         expect(await errorCapturer(client.buildBridgechainResignation({ fee: new BigNumber(fees[2][4] - 1) }))).not.toThrow(/fee/)
         spy.mockRestore()
       })
-    })
-  })
-
-  describe('__signTransaction', () => {
-    const address = Identities.Address.fromPassphrase('passphrase', 23)
-    const publicKey = Identities.PublicKey.fromPassphrase('passphrase')
-    const passphrase = 'passphrase'
-    const secondPassphrase = 'second passphrase'
-    let transaction
-    let signData
-
-    beforeEach(() => {
-      transaction = Transactions.BuilderFactory
-        .transfer()
-        .amount(new BigNumber(1 * 1e8))
-        .fee(new BigNumber(0.1 * 1e8))
-        .recipientId(address)
-        .vendorField('test vendorfield')
-
-      signData = {
-        address,
-        transaction,
-        passphrase,
-        secondPassphrase,
-        networkWif: 170,
-        networkId: 'ark.mainnet'
-      }
-    })
-
-    it('should sign transaction', async () => {
-      setAip11AndSpy(false, false)
-
-      const response = await client.__signTransaction(signData)
-
-      expect(response.vendorField).toEqual(transaction.data.vendorField)
-      expect(response.amount).toBe(new BigNumber(1 * 1e8).toString())
-      expect(response.fee).toBe(new BigNumber(0.1 * 1e8).toString())
-      expect(response.senderPublicKey).toEqual(publicKey)
-      expect(response.type).toEqual(0)
-      expect(response.version).toEqual(1)
-    })
-
-    it('should get network from session if no id', async () => {
-      const networkByIdSpy = jest.spyOn(store.getters, 'network/byId')
-
-      await client.__signTransaction({
-        ...signData,
-        networkId: null
-      })
-
-      expect(networkByIdSpy).not.toHaveBeenCalled()
-
-      networkByIdSpy.mockRestore()
-    })
-
-    it('should get network by id if provided', async () => {
-      const networkByIdSpy = jest.spyOn(store.getters, 'network/byId')
-
-      await client.__signTransaction(signData)
-
-      expect(networkByIdSpy).toHaveBeenCalledWith('ark.mainnet')
-
-      networkByIdSpy.mockRestore()
-    })
-
-    it('should normalize passphrase if provided', async () => {
-      const spy = jest.spyOn(client, 'normalizePassphrase')
-
-      await client.__signTransaction(signData)
-
-      expect(spy).toHaveBeenCalledWith(signData.passphrase)
-      expect(spy).toHaveBeenCalledWith(signData.secondPassphrase)
-    })
-
-    it('should not normalize if no passphrase is provided', async () => {
-      const spy = jest.spyOn(client, 'normalizePassphrase')
-      const wif = Identities.WIF.fromPassphrase(passphrase, { wif: 170 })
-
-      await client.__signTransaction({
-        ...signData,
-        passphrase: null,
-        secondPassphrase: null,
-        wif
-      })
-
-      expect(spy).not.toHaveBeenCalled()
-    })
-
-    it('should create v1 transaction if aip11 disabled', async () => {
-      const response = await client.__signTransaction(signData)
-
-      expect(response.version).toBe(1)
-      expect(response.nonce).toBeFalsy()
-      expect(response.timestamp).toBeTruthy()
-    })
-
-    it('should create v2 transaction if aip11 enabled', async () => {
-      const spy = setAip11AndSpy(true)
-
-      const response = await client.__signTransaction(signData)
-
-      spy.mockRestore()
-
-      expect(response.version).toBe(2)
-      expect(response.nonce).toBe('1')
-    })
-
-    it('should increment nonce of wallet', async () => {
-      const spy = setAip11AndSpy(true)
-      nock('http://127.0.0.1:4003')
-        .get(`/api/v2/wallets/${address}`)
-        .reply(200, {
-          data: {
-            nonce: 3
-          }
-        })
-
-      const response = await client.__signTransaction(signData)
-
-      spy.mockRestore()
-
-      expect(response.version).toBe(2)
-      expect(response.nonce).toBe('4')
-    })
-
-    it('should default nonce to 0 if no wallet', async () => {
-      const spy = setAip11AndSpy(true)
-      nock('http://127.0.0.1:4003')
-        .get(`/api/v2/wallets/${address}`)
-        .reply(200, {
-          data: {}
-        })
-
-      const response = await client.__signTransaction(signData)
-
-      spy.mockRestore()
-
-      expect(response.version).toBe(2)
-      expect(response.nonce).toBe('1')
-    })
-
-    it('should sign with passphrase', async () => {
-      const spy = setAip11AndSpy(true)
-      const spySignPassphrase = jest.spyOn(transaction, 'sign')
-
-      const response = await client.__signTransaction(signData)
-
-      spy.mockRestore()
-
-      expect(response.version).toBe(2)
-      expect(response.nonce).toBe('1')
-      expect(spySignPassphrase).toHaveBeenCalledWith(passphrase)
-    })
-
-    it('should sign with second passphrase', async () => {
-      const spy = setAip11AndSpy(true)
-      const spySecondSignPassphrase = jest.spyOn(transaction, 'secondSign')
-
-      const response = await client.__signTransaction(signData)
-
-      spy.mockRestore()
-
-      expect(response.version).toBe(2)
-      expect(response.nonce).toBe('1')
-      expect(spySecondSignPassphrase).toHaveBeenCalledWith(secondPassphrase)
-    })
-
-    it('should sign with wif', async () => {
-      const spy = setAip11AndSpy(true)
-      const spySignWif = jest.spyOn(transaction, 'signWithWif')
-      const wif = Identities.WIF.fromPassphrase(passphrase, { wif: 170 })
-
-      const response = await client.__signTransaction({
-        ...signData,
-        passphrase: null,
-        secondPassphrase: null,
-        wif
-      })
-
-      spy.mockRestore()
-
-      expect(response.version).toBe(2)
-      expect(response.nonce).toBe('1')
-      expect(spySignWif).toHaveBeenCalledWith(wif, 170)
-    })
-
-    it('should return object', async () => {
-      const spy = setAip11AndSpy(true)
-
-      const response = await client.__signTransaction(signData, true)
-
-      spy.mockRestore()
-
-      expect(response.data).toBeTruthy()
-      expect(response.constructor.name).toBe('TransferBuilder')
-    })
-
-    describe('multiSignature', () => {
-      const minKeys = 3
-      let multiSignature
-      let publicKeys
-      let aip11Spy
-
-      beforeEach(() => {
-        publicKeys = []
-        for (let i = 0; i < 5; i++) {
-          publicKeys.push(Identities.PublicKey.fromPassphrase(`passphrase ${i}`))
-        }
-
-        multiSignature = {
-          publicKeys,
-          min: minKeys
-        }
-        aip11Spy = setAip11AndSpy(true)
-      })
-
-      afterEach(() => {
-        aip11Spy.mockRestore()
-      })
-
-      it('should create transaction for multi-signature wallet when using passphrase', async () => {
-        const getPublicKeyFromPassphrase = WalletService.getPublicKeyFromPassphrase
-        WalletService.getPublicKeyFromPassphrase = jest.fn((passphrase) => Identities.PublicKey.fromPassphrase(passphrase))
-
-        const response = await client.__signTransaction({
-          ...signData,
-          multiSignature
-        })
-
-        WalletService.getPublicKeyFromPassphrase = getPublicKeyFromPassphrase
-
-        expect(response.signatures.length).toBe(0)
-      })
-
-      it('should return transaction with multiSignature property', async () => {
-        const getPublicKeyFromPassphrase = WalletService.getPublicKeyFromPassphrase
-        WalletService.getPublicKeyFromPassphrase = jest.fn((passphrase) => Identities.PublicKey.fromPassphrase(passphrase))
-
-        const response = await client.__signTransaction({
-          ...signData,
-          multiSignature
-        })
-
-        WalletService.getPublicKeyFromPassphrase = getPublicKeyFromPassphrase
-
-        expect(response.multiSignature).toBe(multiSignature)
-      })
-
-      describe('own passphrase used', () => {
-        beforeEach(() => {
-          publicKeys.push(Identities.PublicKey.fromPassphrase(`${passphrase}`))
-        })
-
-        it('should add signature to list of signatures', async () => {
-          const spyMultiSign = jest.spyOn(transaction, 'multiSign')
-          const getPublicKeyFromPassphraseMock = jest.fn((passphrase) => Identities.PublicKey.fromPassphrase(passphrase))
-          const getPublicKeyFromPassphrase = WalletService.getPublicKeyFromPassphrase
-          WalletService.getPublicKeyFromPassphrase = getPublicKeyFromPassphraseMock
-
-          const response = await client.__signTransaction({
-            ...signData,
-            multiSignature
-          })
-
-          WalletService.getPublicKeyFromPassphrase = getPublicKeyFromPassphrase
-
-          const publicKeyIndex = publicKeys.indexOf(publicKey)
-          const signature = response.signatures.find(s => parseInt(s.substring(0, 2), 16) === publicKeyIndex)
-
-          expect(getPublicKeyFromPassphraseMock).toHaveBeenCalledWith(passphrase)
-          expect(spyMultiSign).toHaveBeenCalledWith(passphrase, publicKeyIndex)
-          expect(signature).toBeTruthy()
-        })
-      })
-
-      describe('own wif used', () => {
-        beforeEach(() => {
-          publicKeys.push(Identities.PublicKey.fromPassphrase(`${passphrase}`))
-        })
-
-        it('should add signature to list of signatures', async () => {
-          const spyMultiSignWithWif = jest.spyOn(transaction, 'multiSignWithWif')
-          const getPublicKeyFromWIFMock = jest.fn((wif) => Identities.PublicKey.fromWIF(wif, { wif: 170 }))
-          const getPublicKeyFromWIF = WalletService.getPublicKeyFromWIF
-          WalletService.getPublicKeyFromWIF = getPublicKeyFromWIFMock
-
-          const wif = Identities.WIF.fromPassphrase(passphrase, { wif: 170 })
-          const response = await client.__signTransaction({
-            ...signData,
-            multiSignature,
-            passphrase: null,
-            secondPassphrase: null,
-            wif
-          })
-
-          WalletService.getPublicKeyFromWIF = getPublicKeyFromWIF
-
-          const publicKeyIndex = publicKeys.indexOf(publicKey)
-          const signature = response.signatures.find(s => parseInt(s.substring(0, 2), 16) === publicKeyIndex)
-
-          expect(getPublicKeyFromWIFMock).toHaveBeenCalledWith(wif)
-          expect(spyMultiSignWithWif).toHaveBeenCalledWith(publicKeyIndex, wif, 170)
-          expect(signature).toBeTruthy()
-        })
-      })
-    })
-  })
-
-  describe('multiSign', () => {
-    const masterPassphrase = 'passphrase'
-    const address = Identities.Address.fromPassphrase(masterPassphrase, 23)
-    const publicKey = Identities.PublicKey.fromPassphrase(masterPassphrase)
-    const minKeys = 3
-    let publicKeys
-    let transaction
-    let signData
-    let multiSignature
-    let aip11Spy
-
-    beforeEach(() => {
-      publicKeys = []
-      for (let i = 0; i < 5; i++) {
-        publicKeys.push(Identities.PublicKey.fromPassphrase(`passphrase ${i}`))
-      }
-
-      multiSignature = {
-        publicKeys,
-        min: minKeys
-      }
-      aip11Spy = setAip11AndSpy(true)
-
-      transaction = {
-        amount: new BigNumber(1 * 1e8),
-        fee: new BigNumber(0.1 * 1e8),
-        type: 0,
-        typeGroup: 1,
-        recipientId: address,
-        vendorField: 'test vendorfield',
-        version: 2,
-        network: 23,
-        senderPublicKey: publicKey,
-        nonce: '1',
-        signatures: []
-      }
-
-      signData = {
-        multiSignature,
-        networkWif: 170,
-        passphrase: 'passphrase 1'
-      }
-    })
-
-    afterEach(() => {
-      aip11Spy.mockRestore()
-    })
-
-    it('should throw error if no passphrase or wif', async () => {
-      expect(await errorCapturer(client.multiSign(transaction, { multiSignature }))).toThrow('No passphrase or wif provided')
-    })
-
-    it('should throw error aip11 not enabled', async () => {
-      setAip11AndSpy(false, false)
-
-      expect(await errorCapturer(client.multiSign(transaction, signData))).toThrow('Multi-Signature Transactions are not supported yet')
-    })
-
-    it('should parse transaction from data', async () => {
-      const spy = jest.spyOn(client, '__transactionFromData')
-
-      await client.multiSign(transaction, signData)
-      expect(spy).toHaveBeenCalledWith(transaction)
-    })
-
-    it('should get keys from passphrase if provided', async () => {
-      const spy = jest.spyOn(Identities.Keys, 'fromPassphrase')
-
-      await client.multiSign(transaction, signData)
-      expect(spy).toHaveBeenCalledWith('passphrase 1')
-
-      spy.mockRestore()
-    })
-
-    it('should get keys from wif if provided', async () => {
-      const wif = Identities.WIF.fromPassphrase('passphrase 1', { wif: 170 })
-      const spy = jest.spyOn(Identities.Keys, 'fromWIF')
-
-      await client.multiSign(transaction, {
-        ...signData,
-        passphrase: null,
-        wif
-      })
-      expect(spy).toHaveBeenCalledWith(wif, { wif: 170 })
-
-      spy.mockRestore()
-    })
-
-    it('should check if signatures are needed', async () => {
-      const spy = jest.spyOn(TransactionService, 'isMultiSignatureReady')
-
-      await client.multiSign(transaction, signData)
-
-      expect(spy).toHaveBeenCalledWith({ ...transaction, multiSignature }, true)
-      expect(spy).toHaveBeenCalledTimes(1)
-
-      spy.mockRestore()
-    })
-
-    it('should throw error if passphrase is not required for multi-signature wallet', async () => {
-      expect(await errorCapturer(client.multiSign(transaction, { ...signData, passphrase: 'not used' }))).toThrow('passphrase/wif is not used to sign this transaction')
-    })
-
-    it('should add signature for passphrase', async () => {
-      const response = await client.multiSign(transaction, signData)
-
-      expect(response.signatures.length).toBe(1)
-      expect(response.signatures[0]).toBeTruthy()
-    })
-
-    it('should add additional signatures upto minimum required', async () => {
-      for (let i = 0; i < minKeys; i++) {
-        transaction = await client.multiSign(transaction, { ...signData, passphrase: `passphrase ${i}` })
-      }
-
-      expect(transaction.signatures.length).toBe(3)
-      for (let i = 0; i < minKeys; i++) {
-        const publicKeyIndex = publicKeys.indexOf(Identities.PublicKey.fromPassphrase(`passphrase ${i}`))
-        const signature = transaction.signatures.find(s => parseInt(s.substring(0, 2), 16) === publicKeyIndex)
-
-        expect(signature).toBeTruthy()
-      }
-    })
-
-    it('should not sign transaction if not primary sender', async () => {
-      for (let i = 0; i < 4; i++) {
-        transaction = await client.multiSign(transaction, { ...signData, passphrase: `passphrase ${i}` })
-      }
-
-      expect(transaction.signature).toBeFalsy()
-    })
-
-    it('should ignore duplicate signatures for passphrase', async () => {
-      transaction = await client.multiSign(transaction, { ...signData, passphrase: 'passphrase 1' })
-      transaction = await client.multiSign(transaction, { ...signData, passphrase: 'passphrase 2' })
-      transaction = await client.multiSign(transaction, { ...signData, passphrase: 'passphrase 1' })
-
-      const publicKeyIndex = publicKeys.indexOf(Identities.PublicKey.fromPassphrase('passphrase 1'))
-      const signatures = transaction.signatures.filter(s => parseInt(s.substring(0, 2), 16) === publicKeyIndex)
-      expect(signatures.length).toBe(1)
-    })
-
-    it('should only sign transaction with sender passphrase for registration', async () => {
-      transaction = {
-        fee: new BigNumber(0.1 * 1e8),
-        type: 4,
-        typeGroup: 1,
-        version: 2,
-        network: 23,
-        senderPublicKey: publicKey,
-        nonce: '1',
-        signatures: [],
-        asset: {
-          multiSignature
-        }
-      }
-
-      for (let i = 0; i < 5; i++) {
-        transaction = await client.multiSign(transaction, { ...signData, passphrase: `passphrase ${i}` })
-      }
-      transaction = await client.multiSign(transaction, { ...signData, passphrase: masterPassphrase })
-
-      expect(transaction.signature).toBeTruthy()
-      expect(transaction.signatures.length).toBe(5)
-    })
-
-    it('should sign transaction with sender second passphrase for registration', async () => {
-      transaction = {
-        fee: new BigNumber(0.1 * 1e8),
-        type: 4,
-        typeGroup: 1,
-        version: 2,
-        network: 23,
-        senderPublicKey: publicKey,
-        nonce: '1',
-        signatures: [],
-        asset: {
-          multiSignature
-        }
-      }
-
-      for (let i = 0; i < 5; i++) {
-        transaction = await client.multiSign(transaction, { ...signData, passphrase: `passphrase ${i}` })
-      }
-      transaction = await client.multiSign(transaction, {
-        ...signData,
-        passphrase: masterPassphrase,
-        secondPassphrase: 'second-passphrase'
-      })
-
-      expect(transaction.signature).toBeTruthy()
-      expect(transaction.secondSignature).toBeTruthy()
-      expect(transaction.signatures.length).toBe(5)
-    })
-  })
-
-  describe('__transactionFromData', () => {
-    let transaction
-
-    beforeEach(() => {
-      transaction = {
-        amount: new BigNumber(1 * 1e8),
-        fee: new BigNumber(0.1 * 1e8),
-        type: 0,
-        typeGroup: 1,
-        recipientId: 'address-1',
-        vendorField: 'test vendorfield',
-        version: 2,
-        network: 23,
-        senderPublicKey: 'publicKey-1',
-        timestamp: 100000,
-        nonce: '1',
-        signatures: [],
-        multiSignature: {
-          min: 3,
-          publicKeys: [
-            1,
-            2,
-            3
-          ]
-        }
-      }
-    })
-
-    it('should do a deep clone', () => {
-      const clonedTransaction = client.__transactionFromData(transaction)
-      transaction.amount = new BigNumber(2 * 1e8)
-
-      expect(clonedTransaction.amount + '').toEqual('100000000')
-    })
-
-    it('should remove unnecessary properties', () => {
-      const clonedTransaction = client.__transactionFromData(transaction)
-
-      expect(clonedTransaction.timestamp).toBe(undefined)
-      expect(clonedTransaction.multiSignature).toBe(undefined)
     })
   })
