+++ src/renderer/services/crypto/business-update.builder.js
@@ -2,7 +2,7 @@ import * as MagistrateCrypto from '@arkecosystem/core-magistrate-crypto'
 import { TRANSACTION_TYPES } from '@config'
 import store from '@/store'
 import { CryptoUtils } from './utils'
-import TransactionSigner from './transaction-signer'
+import { TransactionSigner } from './transaction-signer'
 
 export class BusinessUpdateBuilder {
   static async build (
