+++ src/renderer/services/crypto/delegate-resignation.builder.js
@@ -2,7 +2,7 @@ import { Transactions } from '@arkecosystem/crypto'
 import { TRANSACTION_TYPES } from '@config'
 import store from '@/store'
 import { CryptoUtils } from './utils'
-import TransactionSigner from './transaction-signer'
+import { TransactionSigner } from './transaction-signer'
 
 export class DelegateResignationBuilder {
   static async build (
