+++ src/renderer/services/crypto/multi-payment.builder.js
@@ -2,7 +2,7 @@ import { Transactions } from '@arkecosystem/crypto'
 import { TRANSACTION_TYPES } from '@config'
 import store from '@/store'
 import { CryptoUtils } from './utils'
-import TransactionSigner from './transaction-signer'
+import { TransactionSigner } from './transaction-signer'
 
 export class MultiPaymentBuilder {
   static async build (
